+++ robotrace_v1/Core/Src/lineTrace.c
@@ -7,7 +7,7 @@
 //====================================//
 int16_t		tracePwm;		// 白線トレースサーボPWM
 int16_t 	traceBefore;	// 1ms前のセンサ値
-double		Int;			// I成分積算値(白線トレース)
+float		Int;			// I成分積算値(白線トレース)
 uint8_t		kp1_buff = KP1, ki1_buff = KI1, kd1_buff = KD1;
 uint8_t		kp1Curve_buff = KP1CURVE, ki1Curve_buff = KI1CURVE, kd1Curve_buff = KD1CURVE;
 
@@ -43,13 +43,13 @@ void motorControlTrace( void ) {
 	// kd = kd1_buff;
 	// Dev = angleSensor * 10;
 	// I成分積算
-	Int += (double)Dev * 0.001;
+	Int += (float)Dev * 0.001;
 	if ( Int > 10000 ) Int = 10000;		// I成分リミット
 	else if ( Int < -10000 ) Int = -10000;
 	Dif = ( Dev - traceBefore ) * 1;	// dゲイン1/1000倍
 
 	iP = (int32_t)kp1_buff * Dev;	// 比例
-	iI = (double)ki1_buff * Int;	// 積分
+	iI = (float)ki1_buff * Int;	// 積分
 	iD = (int32_t)kd1_buff * Dif;	// 微分
 	iRet = iP + iI + iD;
 	iRet = iRet >> 6;				// PWMを0～100近傍に収める
