+++ robotrace_v1/Core/Src/yawRateCtrl.c
@@ -5,10 +5,10 @@
 //====================================//
 // グローバル変数の宣言
 //====================================//
-double	 	targetAngularVelocity;
+float	 	targetAngularVelocity;
 int32_t		yawRatePwm;				// 制御量
 int32_t 	angularVelocityBefore;	// 1ms前のセンサ値
-double		Int3;					// I成分積算値(白線トレース)
+float		Int3;					// I成分積算値(白線トレース)
 uint8_t		kp3_buff = KP3, ki3_buff = KI3, kd3_buff = KD3;
 
 ///////////////////////////////////////////////////////////////////////////
@@ -18,7 +18,7 @@ uint8_t		kp3_buff = KP3, ki3_buff = KI3, kd3_buff = KD3;
 // 戻り値       なし
 ///////////////////////////////////////////////////////////////////////////
 void motorControlYawRate(void) {
-	double iP, iD, iI;
+	float iP, iD, iI;
 	int32_t Dev, Dif, kp, ki, kd, iRet;
 	
 	//サーボモータ用PWM値計算
@@ -27,14 +27,14 @@ void motorControlYawRate(void) {
 	kd = kd3_buff;
 	Dev = angularVelocity[INDEX_Z] - targetAngularVelocity;	// 目標値-現在値
 	// I成分積算
-	Int3 += (double)Dev * 0.005;
+	Int3 += (float)Dev * 0.005;
 	// if ( Int3 > 5000 ) Int3 = 5000;		// I成分リミット
 	// else if ( Int3 < -5000 ) Int3 = -5000;s
 	Dif = ( Dev - angularVelocityBefore ) * 1;	// dゲイン1/1000倍
 
-	iP = (double)kp * Dev;	// 比例
-	iI = (double)ki * Int3;	// 積分
-	iD = (double)kd * Dif;	// 微分
+	iP = (float)kp * Dev;	// 比例
+	iI = (float)ki * Int3;	// 積分
+	iD = (float)kd * Dif;	// 微分
 	iRet = (int32_t)iP + iI + iD;
 	iRet = iRet >> 2;				// PWMを0～100近傍に収める
